backend/app/schemas/todo.py
from pydantic import BaseModel
from datetime import datetime
from typing import Optional

class TodoBase(BaseModel):
    title: str
    is_completed: bool = False

class TodoCreate(TodoBase):
    pass

class TodoUpdate(BaseModel):
    title: Optional[str] = None
    is_completed: Optional[bool] = None

class Todo(TodoBase):
    id: int
    user_id: int
    created_at: datetime

    class Config:
        from_attributes = True
