backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-dotenv
gunicorn
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
